migrate the fetch call from Main to SearchResults
