+++ src/Components/Main.jsx
@@ -1,31 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { Button } from 'react-bootstrap'
-import axios from 'axios';
 
 function Main(props) {
 
     const [search, setSearch] = useState('')
-    const [profiles, setProfiles] = useState([])
-
-    useEffect(() => {
-        getAllProfiles()
-    }, [])
-
-    const getAllProfiles = async () => {
-        console.log("Attempting to retrieve all profiles...")
-
-        try {
-            const url =
-                process.env.NODE_ENV === 'production'
-                    ? `https://deployed-heroku-api.com/profiles`
-                    : `http://localhost:4000/profiles`
-
-            const allProfiles = await axios(url)
-            setProfiles(allProfiles.data)
-        } catch (error) {
-            console.warn("API call isn't fully set up yet.")
-        }
-    }
 
     return (
         <div className="searchMain">
